AND not OR

+++ static/build/.tmp_update/script/blue_light.sh
@@ -168,7 +168,7 @@ disable_blue_light_filter() {
 
 check_blf() {
     sync
-    if [ -f "$blf_key" ] || [ ! -f "$ignore_schedule" ]; then
+    if [ -f "$blf_key" ] && [ ! -f "$ignore_schedule" ]; then
 
         if [ ! -f "$blf_key_on" ] && [ -f "$blf_key_on_user" ]; then
             rm -f $blf_key_on_user
